finscope-backend/finscope-backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
httpx==0.27.0
pydantic==2.8.2
